Split tests into separate files
 ## Summary
pytest/fixtures/selenium don't like it when they run multiple functional
tests declared in a single file... So we'll try splitting them into
separate files. We should probably review out functional test setup
later in a tech debt ticket.
